refactor replay tests
